requirements.txt
numpy==1.26.3
scipy==1.10.1
sentence-transformers==2.7.0
torchvision==0.19.1
gensim==4.3.3
scikit-learn==1.5.1
tqdm==4.66.5
wandb==0.18.1
topmost==0.0.5
geomloss==0.2.6
